Add files via upload
The Bank Loan Approval Prediction project aims to predict whether a loan application will be approved based on various financial and personal features of the applicant, such as income, credit score, loan amount, and more. The model is built using Logistic Regression, a binary classification algorithm, to classify loan applications as either approved (1) or rejected (0).

To optimize the performance of the logistic regression model, Grid Search Cross-Validation (GridSearchCV) is applied for hyperparameter tuning. This allows for an exhaustive search over specified parameter values to find the best combination, leading to improved model accuracy and robustness.